Move PublicVisibility to validation package
