+++ workflow-service/src/main/java/com/redhat/parodos/workflow/execution/validation/PublicVisibilityValidator.java
@@ -1,4 +1,4 @@
-package com.redhat.parodos.workflow.execution.controller;
+package com.redhat.parodos.workflow.execution.validation;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
