+++ workflow-service/src/test/java/com/redhat/parodos/workflow/execution/validation/PubliclyVisibleValidatorTest.java
@@ -1,4 +1,4 @@
-package com.redhat.parodos.workflow.execution.controller;
+package com.redhat.parodos.workflow.execution.validation;
 
 import javax.validation.ConstraintValidatorContext;
